refactor(runtime): factor out `useControlDefs` hook
